Change package name: plsa -> topicmodel
