+++ core/src/test/java/hivemall/topicmodel/PLSAUDTFTest.java
@@ -16,46 +16,48 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package hivemall.plsa;
+package hivemall.topicmodel;
 
-import java.util.Map;
 import java.util.List;
+import java.util.Map;
 import java.util.SortedMap;
+import java.util.Arrays;
+
+import org.apache.hadoop.hive.ql.metadata.HiveException;
+import org.apache.hadoop.hive.serde2.objectinspector.ObjectInspector;
+import org.apache.hadoop.hive.serde2.objectinspector.ObjectInspectorFactory;
+import org.apache.hadoop.hive.serde2.objectinspector.ObjectInspectorUtils;
+import org.apache.hadoop.hive.serde2.objectinspector.primitive.PrimitiveObjectInspectorFactory;
 
 import org.junit.Assert;
 import org.junit.Test;
 
-public class IncrementalPLSAModelTest {
+public class PLSAUDTFTest {
     private static final boolean DEBUG = false;
 
     @Test
-    public void test() {
-        int K = 2;
-        int it = 0;
-        float perplexityPrev;
-        float perplexity = Float.MAX_VALUE;
+    public void test() throws HiveException {
+        PLSAUDTF udtf = new PLSAUDTF();
 
-        IncrementalPLSAModel model = new IncrementalPLSAModel(K, 1E-5f, 1E-5d);
+        ObjectInspector[] argOIs = new ObjectInspector[] {
+                ObjectInspectorFactory.getStandardListObjectInspector(PrimitiveObjectInspectorFactory.javaStringObjectInspector),
+                ObjectInspectorUtils.getConstantObjectInspector(
+                    PrimitiveObjectInspectorFactory.javaStringObjectInspector, "-topic 2 -alpha 0.00001 -delta 0.00001")};
 
-        String[] doc1 = new String[] {"fruits:1", "healthy:1", "vegetables:1"};
-        String[] doc2 = new String[] {"apples:1", "avocados:1", "colds:1", "flu:1", "like:2", "oranges:1"};
-        String[][] miniBatch = new String[][] {doc1, doc2};
+        udtf.initialize(argOIs);
 
-        do {
-            // online (i.e., one-by-one) updating
-            model.train(miniBatch);
-
-            it++;
-            perplexityPrev = perplexity;
-            perplexity = model.computePerplexity();
-            println("Iteration " + it + ": perplexity = " + perplexity);
-        } while(Math.abs(perplexityPrev - perplexity) >= 1E-5f);
+        String[] doc1 = new String[]{"fruits:1", "healthy:1", "vegetables:1"};
+        String[] doc2 = new String[]{"apples:1", "avocados:1", "colds:1", "flu:1", "like:2", "oranges:1"};
+        for (int it = 0; it < 10000; it++) {
+            udtf.process(new Object[]{ Arrays.asList(doc1) });
+            udtf.process(new Object[]{ Arrays.asList(doc2) });
+        }
 
         SortedMap<Float, List<String>> topicWords;
 
         println("Topic 0:");
         println("========");
-        topicWords = model.getTopicWords(0);
+        topicWords = udtf.getTopicWords(0);
         for (Map.Entry<Float, List<String>> e : topicWords.entrySet()) {
             List<String> words = e.getValue();
             for (int i = 0; i < words.size(); i++) {
@@ -66,7 +68,7 @@ public class IncrementalPLSAModelTest {
 
         println("Topic 1:");
         println("========");
-        topicWords = model.getTopicWords(1);
+        topicWords = udtf.getTopicWords(1);
         for (Map.Entry<Float, List<String>> e : topicWords.entrySet()) {
             List<String> words = e.getValue();
             for (int i = 0; i < words.size(); i++) {
@@ -75,9 +77,8 @@ public class IncrementalPLSAModelTest {
         }
         println("========");
 
-
         int k1, k2;
-        float[] topicDistr = model.getTopicDistribution(doc1);
+        float[] topicDistr = udtf.getTopicDistribution(doc1);
         if (topicDistr[0] > topicDistr[1]) {
             // topic 0 MUST represent doc#1
             k1 = 0;
@@ -86,12 +87,13 @@ public class IncrementalPLSAModelTest {
             k1 = 1;
             k2 = 0;
         }
+
         Assert.assertTrue("doc1 is in topic " + k1 + " (" + (topicDistr[k1] * 100) + "%), "
             + "and `vegetables` SHOULD be more suitable topic word than `flu` in the topic",
-            model.getProbability("vegetables", k1) > model.getProbability("flu", k1));
+            udtf.getProbability("vegetables", k1) > udtf.getProbability("flu", k1));
         Assert.assertTrue("doc2 is in topic " + k2 + " (" + (topicDistr[k2] * 100) + "%), "
             + "and `avocados` SHOULD be more suitable topic word than `healthy` in the topic",
-            model.getProbability("avocados", k2) > model.getProbability("healthy", k2));
+            udtf.getProbability("avocados", k2) > udtf.getProbability("healthy", k2));
     }
 
     private static void println(String msg) {
