+++ resources/ddl/define-all-as-permanent.hive
@@ -627,10 +627,10 @@ DROP FUNCTION IF EXISTS lda_predict;
 CREATE FUNCTION lda_predict as 'hivemall.topicmodel.LDAPredictUDAF' USING JAR '${hivemall_jar}';
 
 DROP FUNCTION IF EXISTS train_plsa;
-CREATE FUNCTION train_plsa as 'hivemall.plsa.PLSAUDTF' USING JAR '${hivemall_jar}';
+CREATE FUNCTION train_plsa as 'hivemall.topicmodel.PLSAUDTF' USING JAR '${hivemall_jar}';
 
 DROP FUNCTION IF EXISTS plsa_predict;
-CREATE FUNCTION plsa_predict as 'hivemall.plsa.PLSAPredictUDAF' USING JAR '${hivemall_jar}';
+CREATE FUNCTION plsa_predict as 'hivemall.topicmodel.PLSAPredictUDAF' USING JAR '${hivemall_jar}';
 
 ---------------------------
 -- Geo-Spatial functions --
